requirements.txt
# local package
-e .

# external requirements
torch
transformers
tqdm
numpy
pandas
scikit-learn
wandb
scipy
soundfile
python-dotenv>=0.5.1
